made home page and created hotel plus review feature(idk if it works)
so i added a hotel page with a review feature. idk if it works.  i also added some changes to the home /index page. like a flex navbar. i have a controller page, a model page, a service page, and a repository page. i might need to rename so that it doesnt clash with other files.